Move pandigital() to tools.
